Updates to author blocks (so they don't drown out the rest in the documentation and removed borg.py
borg.py has been replaced by data_assimilation.py and the author blocks as they were were absolutely
giant in the Sphinx docs. The emphasis should be on the code, not who wrote it, so I made the block
just a couple of lines of text without using a ReST header (======).