datamodel/test_datamodel: fixed alteration of datetimes in datamodel on write to HDF5. Updated test.
One of the test updates removed the datamodel test for datetime<->HDF5. There has never been functionality to read an ISO time from an HDF5 and convert automatically to datetime. While we may want to implement this, it should be filed as a feature request (if wanted) and then an implementation needs to thought through. Liekly, adding a SPACEPY_CONVERT attribute, or similar, then trapping this on load, would be a good solution.