empiricals: corrected Vampola model to account for 4.pi difference between integrating directional flux over pitch angle and omnidirectional flux.
The reasoning for this is that omnidirectional flux covers 4.pi steradians of the sky, whereas
directional flux over 0 to 180 degrees (-pi/2 to pi/2 radians) does not cover all look angles.
The integral over look angles implies a factor of 2.pi and only integrating over 0-90 degrees,
(as in the Vampola model where we assume symmetry in the pitch angle distribution) implies
another factor of 2. Thus, a factor of 4.pi.