tests/test_toolbox: corrected windowMean tests assuming window is exclusive and windows cover full data range
This change in window exclusivity assumption is necessary to ensure that we don't double count points
that we don't mean to. The data range change is to ensure that we average all data present, and
don't lose information. With an overlap of zero adjacent windows shouldn't share any common points. The
changed logic checks for that.