datamodel: fixed bug in reading global attrs where non-string values would fail
Issue reported by DTW. Code would fail on ints, bools, etc. stored in global attrs
as it would check for presence of START_COLUMN to determine whether it was a variable
or a global attr. Fixed to now check for keys attribute first so we aren't
inappropriately testing attributes. Added a test to catch the previous bad behaviour.