pybats: trap MemoryError reading ascii IDL file with binary assumption
Type guessing on Windows leads to a MemoryError when the ASCII file is
read using _read_idl_bin. This traps for the error, and also wraps the
whole _read_idl_bin in a context manager so that the file is closed on
raising any exception. A test is added to make sure the typeguessing
doesn't fail. Fixes #86.