coordinates: reconcile Re definition for astropy merge
Updating after rebasing some big changes from the master branch
This includes the to/from astropy, which moved the Earth radius
definition and changed it from km to metres. This commit changes
it back to metres, and enforces the conversion km->m in the
astropy routines. It's not as clean, but we actually use the Re
definition now and this minimizes the change to the new code.