Added sort_unstructured_data keyword
Added the additional optional sort_unstructured keyword in a way that preserves the default behavior of IdlFile by setting the default to None.
If the keyword remains None when calling the read() method of IdlFile, the keyword will be set to True when reading a binary file via _read_idl_bin() and to False when reading an ASCII file via _read_idl_ascii().
Otherwise the Boolean True or False value entered by the user when instantiating the IdlFile class will be transmitted to the appropriate reader.

Sorting unstructured data to place nearby positions close to each other (rather than preserve arrangement within (slices of) blocks in the BATSRUS grid) was done when reading binary files but not when reading ASCII files (the reader did not implement the method). T
The preservation of the grid structure in the imported data is needed when using the block structure in further processing of 3D files. 
The Kamodo (github.com/nasa/kamodo) model access, analysis and visualization package employs spacepy.bybats to read 3D BATSRUS magnetosphere output files and implements an interpolator that used the grid structure and thus can avoid the otherwise necessary but computationally expensive Delauney triangulation available in Scipy interpolation methods.